backend/requirements.txt
django>=5.0
djangorestframework
django-cors-headers
firebase-admin
python-dotenv
anthropic
pdfplumber
boto3
